Meena-based generator Context: Making a generator using the meena model (go/meena) to generate counterfactuals.
We want to enable the user to make on-the-fly text transform rules for a sentence or sentences.

In standard few-shot learning, if a user would wanted to have a transform rule (ie, "make it shorter") they would give a few examples of texts, along with shortened versions of those texts.

This is slightly different, and  uses "meta" few shot learning. That is, we set up a context with Meena with examples that give one-shot learned suggestions (ie, make a text shorter or happier), and Meena learns the meta task of transforming an arbitrary sentence given an arbitrary command.

PiperOrigin-RevId: 351362620